Finish Line
I put in a temporary finish line that slowly moves down. It has a
function with a variable time that allows you to stop it for that amount
of time. This function is currently called when a player runs into an
enemy. There are two finish lines, each attached to their respected
players.